Reinstate an Html module
So, I removed the old HTML module earlier in this v2 drive because I kind of
hated it. It seemed gratuitious. But I later realised I needed to keep a cleanish
interface like this in order to support the extension gems. So it's back. Sort
of. I've renamed the module to a) match the Formtastic module and b) stop fighting
Ruby's camel case expectations. So the extension gems will still need an upgrade.