Additional explanation for error messages
Changed the rails docs link to a more specific anchor and added an example of what to change and where to change it. I'm fairly new to Rails, and had to do quite a bit of digging before I found how to properly structure my config/locales/en.yml file to override error messages for Judge. This edit should save some time for others new to Rails' i18n structure.